+ Updated entire pod repo!!

+++ lib/ui/AsyncImageView.m
@@ -131,7 +131,7 @@ delegate=nil;
 
 	if ([response respondsToSelector:@selector(statusCode)])
 	{
-		int statusCode = [((NSHTTPURLResponse *)response) statusCode];
+		NSInteger statusCode = [((NSHTTPURLResponse *)response) statusCode];
 		if (statusCode >= 400)
 		{
 			[connection cancel];  // stop connecting; no more delegate messages
